refactor(benchmark): use ValueTotalSize for both valueContains args
Change valueContains benchmark to use ValueTotalSize wrapper for both container and contained arguments, instead of ValueLogOuterSizeAddLogMaxInnerSize for the container. This addresses Kenneth's observation that using total size for both will straighten out the curved edge of the benchmark data and make the early-exit boundary at n₁ = n₂ visible as a straight line.